+++ trikControl/src/guiWorker.h
@@ -50,7 +50,7 @@ public slots:
 	/// for better performance.
 	void showImage(const QString &fileName);
 
-	void show(const QVector<uint8_t> &array, int width, int height, const QString &format);
+	void show(const QVector<int32_t> &array, int width, int height, const QString &format);
 
 	/// Add a label to the specific position of the screen without redrawing it.
 	/// If there already is a label in these coordinates, its contents will be updated.
